Add admin protected routing (#55)
* Add admin protected routing

* v0.9.0

* Add secrects

+++ packages/service/src/tests/integretion/controllers/categories.test.ts
@@ -5,7 +5,12 @@ import { Server } from 'http';
 import appStartUp from '../../../app';
 import Category from '../../../models/category';
 import User from '../../../models/user';
-import { signUpUser, createCategories, getCategories } from '../../../utils/test-basic-calls';
+import {
+  signUpUser,
+  createCategories,
+  getCategories,
+  loginAdmin,
+} from '../../../utils/test-basic-calls';
 
 const endpoint = '/api/categories/';
 
@@ -18,7 +23,7 @@ describe('Controller: Category', () => {
 
   afterEach(async () => {
     await Category.deleteMany({});
-    await User.deleteMany({});
+    await User.deleteMany({ name: { $ne: 'eportal_admin' } });
   });
 
   afterAll(async () => {
@@ -28,8 +33,8 @@ describe('Controller: Category', () => {
 
   describe('createCategories Controller', () => {
     it('should create category', async () => {
-      const user = await signUpUser(server);
-      const categories = await createCategories(server, user.body.accessToken);
+      const admin = await loginAdmin(server);
+      const categories = await createCategories(server, admin.body.accessToken);
 
       expect(categories.status).toEqual(201);
       expect(categories.body.categories).toBeDefined();
@@ -40,6 +45,13 @@ describe('Controller: Category', () => {
       expect(categories.body.categories[0]._id).toBeDefined();
       expect(categories.body.categories[0].date).toBeDefined();
       expect(categories.body.categories[0].children).toBeDefined();
+    });
+
+    it('should NOT create category if no admin', async () => {
+      const user = await signUpUser(server);
+      const categories = await createCategories(server, user.body.accessToken);
+
+      expect(categories.status).toEqual(401);
     });
 
     it('should NOT return userData if access token not provided', async () => {
@@ -56,12 +68,12 @@ describe('Controller: Category', () => {
     });
 
     it('should return 422 if body is not valid', async () => {
-      const user = await signUpUser(server);
-      const withOutNames = await createCategories(server, user.body.accessToken, {
+      const admin = await loginAdmin(server);
+      const withOutNames = await createCategories(server, admin.body.accessToken, {
         categories: [{ description: 'description1' }, { name: '', description: 'description2' }],
       });
 
-      const emptyCategories = await createCategories(server, user.body.accessToken, {
+      const emptyCategories = await createCategories(server, admin.body.accessToken, {
         categories: [{ description: 'description1' }, { name: '', description: 'description2' }],
       });
 
@@ -70,10 +82,10 @@ describe('Controller: Category', () => {
     });
 
     it('should update parent categories with new childs', async () => {
-      const user = await signUpUser(server);
-      const createdCategories = await createCategories(server, user.body.accessToken);
+      const admin = await loginAdmin(server);
+      const createdCategories = await createCategories(server, admin.body.accessToken);
 
-      const childCategory = await createCategories(server, user.body.accessToken, {
+      const childCategory = await createCategories(server, admin.body.accessToken, {
         categories: [
           {
             name: 'cat no3',
@@ -83,7 +95,7 @@ describe('Controller: Category', () => {
         ],
       });
 
-      const categories = await getCategories(server, user.body.accessToken);
+      const categories = await getCategories(server, admin.body.accessToken);
       expect(categories.body.categories).toHaveLength(3);
       expect(categories.body.categories[1].children).toHaveLength(1);
       expect(categories.body.categories[1].children[0]._id).toEqual(
@@ -94,16 +106,20 @@ describe('Controller: Category', () => {
 
   describe('getCategories Controller', () => {
     it('should get all created categories', async () => {
+      const admin = await loginAdmin(server);
+      await createCategories(server, admin.body.accessToken);
+
       const user = await signUpUser(server);
-      await createCategories(server, user.body.accessToken);
 
       const categories = await getCategories(server, user.body.accessToken);
       expect(categories.body.categories).toHaveLength(2);
     });
 
     it('should get only selected category', async () => {
+      const admin = await loginAdmin(server);
+      const createdCats = await createCategories(server, admin.body.accessToken);
+
       const user = await signUpUser(server);
-      const createdCats = await createCategories(server, user.body.accessToken);
 
       const categories = await request(server)
         .get('/api/categories?ids=' + createdCats.body.categories[1]._id)
