+++ packages/service/src/tests/integretion/controllers/pages.test.ts
@@ -4,7 +4,13 @@ import { Server } from 'http';
 import appStartUp from '../../../app';
 import Page from '../../../models/page';
 import User from '../../../models/user';
-import { signUpUser, createPage, getPage, updatePage } from '../../../utils/test-basic-calls';
+import {
+  signUpUser,
+  createPage,
+  getPage,
+  updatePage,
+  loginAdmin,
+} from '../../../utils/test-basic-calls';
 
 describe('Controller: Page', () => {
   let server: Server;
@@ -15,7 +21,7 @@ describe('Controller: Page', () => {
 
   afterEach(async () => {
     await Page.deleteMany({});
-    await User.deleteMany({});
+    await User.deleteMany({ name: { $ne: 'eportal_admin' } });
   });
 
   afterAll(async () => {
@@ -25,14 +31,21 @@ describe('Controller: Page', () => {
 
   describe('createPage Controller', () => {
     it('should create page', async () => {
-      const user = await signUpUser(server);
-      const page = await createPage(server, user.body.accessToken, user.body.userData.id);
+      const admin = await loginAdmin(server);
+      const page = await createPage(server, admin.body.accessToken, admin.body.userData.id);
 
       expect(page.status).toEqual(201);
       expect(page.body.page).toBeDefined();
       expect(page.body.page.content.title).toEqual('TITLE');
       expect(page.body.page.content.content).toEqual('CONTENT');
       expect(page.body.page.slug).toEqual('title');
+    });
+
+    it('should NOT create page if not admin', async () => {
+      const user = await signUpUser(server);
+      const page = await createPage(server, user.body.accessToken, user.body.userData.id);
+
+      expect(page.status).toEqual(401);
     });
 
     it('should NOT create page if access token not provided', async () => {
@@ -42,15 +55,15 @@ describe('Controller: Page', () => {
     });
 
     it('should return 422 if body is not valid', async () => {
-      const user = await signUpUser(server);
-      const withoutAuthor = await createPage(server, user.body.accessToken, '');
-      const withoutContent = await createPage(server, user.body.accessToken, '', {
+      const admin = await loginAdmin(server);
+      const withoutAuthor = await createPage(server, admin.body.accessToken, '');
+      const withoutContent = await createPage(server, admin.body.accessToken, '', {
         content: {},
-        author: user.body.userData.id,
+        author: admin.body.userData.id,
       });
-      const withoutEmptyTitle = await createPage(server, user.body.accessToken, '', {
+      const withoutEmptyTitle = await createPage(server, admin.body.accessToken, '', {
         content: { title: '', content: 'CONTENT' },
-        author: user.body.userData.id,
+        author: admin.body.userData.id,
       });
 
       expect(withoutAuthor.status).toEqual(422);
@@ -61,8 +74,10 @@ describe('Controller: Page', () => {
 
   describe('getPage Controller', () => {
     it('should get page', async () => {
+      const admin = await loginAdmin(server);
+      await createPage(server, admin.body.accessToken, admin.body.userData.id);
+
       const user = await signUpUser(server);
-      await createPage(server, user.body.accessToken, user.body.userData.id);
       const page = await getPage(server, user.body.accessToken, 'title');
 
       expect(page.status).toEqual(200);
@@ -73,8 +88,8 @@ describe('Controller: Page', () => {
     });
 
     it('should NOT return page data if access token not provided', async () => {
-      const user = await signUpUser(server);
-      await createPage(server, user.body.accessToken, user.body.userData.id);
+      const admin = await loginAdmin(server);
+      await createPage(server, admin.body.accessToken, admin.body.userData.id);
       const page = await getPage(server, 'BAD-TOKEN', 'title');
 
       expect(page.status).toEqual(401);
@@ -83,11 +98,11 @@ describe('Controller: Page', () => {
 
   describe('updatePage Controller', () => {
     it('should update page', async () => {
-      const user = await signUpUser(server);
-      const createdPage = await createPage(server, user.body.accessToken, user.body.userData.id);
+      const admin = await loginAdmin(server);
+      const createdPage = await createPage(server, admin.body.accessToken, admin.body.userData.id);
       const updatedPage = await updatePage(
         server,
-        user.body.accessToken,
+        admin.body.accessToken,
         createdPage.body.page.slug,
       );
 
@@ -98,9 +113,23 @@ describe('Controller: Page', () => {
       expect(updatedPage.body.page.slug).toEqual('title');
     });
 
-    it('should NOT update page if access token not provided', async () => {
+    it('should NOT update page if not admin', async () => {
+      const admin = await loginAdmin(server);
+      const createdPage = await createPage(server, admin.body.accessToken, admin.body.userData.id);
+
       const user = await signUpUser(server);
-      const createdPage = await createPage(server, user.body.accessToken, user.body.userData.id);
+      const updatedPage = await updatePage(
+        server,
+        user.body.accessToken,
+        createdPage.body.page.slug,
+      );
+
+      expect(updatedPage.status).toEqual(401);
+    });
+
+    it('should NOT update page if access token not provided', async () => {
+      const admin = await loginAdmin(server);
+      const createdPage = await createPage(server, admin.body.accessToken, admin.body.userData.id);
       const updatedPage = await updatePage(server, 'BAD-TOKEN', createdPage.body.page.slug);
 
       expect(updatedPage.status).toEqual(401);
